Move graphql_controller to API v1
